requirements.txt
pytest
python-coveralls
pytest-cov
python-dotenv
flask
flask-sqlalchemy
flask-migrate
gunicorn
graphene
graphene_sqlalchemy
flask_graphql
psycopg2

